# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal Budget Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: accounts
    description: Budget accounts
  - name: transactions
    description: Financial transactions
  - name: categories
    description: Transaction categories
  - name: budgets
    description: Budget limits per category
  - name: summary
    description: Dashboard summaries
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"

  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions with optional filters
      parameters:
        - name: account_id
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: category_id
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          required: false
          schema:
            type: ["string", "null"]
            enum: [income, expense, null]
        - name: month
          in: query
          required: false
          description: "Filter by month (YYYY-MM)"
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: List all budgets
      parameters:
        - name: month
          in: query
          required: false
          description: "Filter by month (YYYY-MM)"
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create or update a budget for a category/month
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"

  /budgets/{id}:
    patch:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a budget limit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetUpdate"
      responses:
        "200":
          description: Updated budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /summary/overview:
    get:
      operationId: getSummaryOverview
      tags: [summary]
      summary: Dashboard overview — total balance, income, expenses for current month
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Overview summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverviewSummary"

  /summary/spending-by-category:
    get:
      operationId: getSpendingByCategory
      tags: [summary]
      summary: Spending breakdown by category for a given month
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Spending by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySpending"

  /summary/monthly-trend:
    get:
      operationId: getMonthlyTrend
      tags: [summary]
      summary: Monthly income vs expense trend for the last 6 months
      responses:
        "200":
          description: Monthly trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyTrend"

  /summary/budget-vs-actual:
    get:
      operationId: getBudgetVsActual
      tags: [summary]
      summary: Budget vs actual spending per category for a given month
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Budget vs actual
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetVsActual"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Account:
      type: object
      required: [id, name, type, balance, currency, created_at]
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string, enum: [checking, savings, credit, cash, investment] }
        balance: { type: number }
        currency: { type: string }
        color: { type: ["string", "null"] }
        created_at: { type: string, format: date-time }

    AccountInput:
      type: object
      required: [name, type, balance]
      properties:
        name: { type: string, minLength: 1 }
        type: { type: string, enum: [checking, savings, credit, cash, investment] }
        balance: { type: number }
        currency: { type: string, default: USD }
        color: { type: string }

    AccountUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        type: { type: string, enum: [checking, savings, credit, cash, investment] }
        balance: { type: number }
        color: { type: string }

    Category:
      type: object
      required: [id, name, type, color, icon]
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string, enum: [income, expense] }
        color: { type: string }
        icon: { type: string }

    CategoryInput:
      type: object
      required: [name, type]
      properties:
        name: { type: string, minLength: 1 }
        type: { type: string, enum: [income, expense] }
        color: { type: string, default: "#6366f1" }
        icon: { type: string, default: "circle" }

    CategoryUpdate:
      type: object
      properties:
        name: { type: string }
        color: { type: string }
        icon: { type: string }

    Transaction:
      type: object
      required: [id, account_id, category_id, amount, type, description, date, created_at]
      properties:
        id: { type: integer }
        account_id: { type: integer }
        category_id: { type: integer }
        account_name: { type: ["string", "null"] }
        category_name: { type: ["string", "null"] }
        category_color: { type: ["string", "null"] }
        category_icon: { type: ["string", "null"] }
        amount: { type: number }
        type: { type: string, enum: [income, expense] }
        description: { type: string }
        notes: { type: ["string", "null"] }
        date: { type: string, format: date }
        created_at: { type: string, format: date-time }

    TransactionInput:
      type: object
      required: [account_id, category_id, amount, type, description, date]
      properties:
        account_id: { type: integer }
        category_id: { type: integer }
        amount: { type: number, minimum: 0.01 }
        type: { type: string, enum: [income, expense] }
        description: { type: string, minLength: 1 }
        notes: { type: string }
        date: { type: string, format: date }

    TransactionUpdate:
      type: object
      properties:
        account_id: { type: integer }
        category_id: { type: integer }
        amount: { type: number, minimum: 0.01 }
        type: { type: string, enum: [income, expense] }
        description: { type: string }
        notes: { type: string }
        date: { type: string, format: date }

    Budget:
      type: object
      required: [id, category_id, month, limit_amount, spent_amount]
      properties:
        id: { type: integer }
        category_id: { type: integer }
        category_name: { type: ["string", "null"] }
        category_color: { type: ["string", "null"] }
        category_icon: { type: ["string", "null"] }
        month: { type: string, description: "YYYY-MM format" }
        limit_amount: { type: number }
        spent_amount: { type: number }

    BudgetInput:
      type: object
      required: [category_id, month, limit_amount]
      properties:
        category_id: { type: integer }
        month: { type: string }
        limit_amount: { type: number, minimum: 0.01 }

    BudgetUpdate:
      type: object
      properties:
        limit_amount: { type: number, minimum: 0.01 }

    OverviewSummary:
      type: object
      required: [total_balance, monthly_income, monthly_expenses, net_savings, transaction_count]
      properties:
        total_balance: { type: number }
        monthly_income: { type: number }
        monthly_expenses: { type: number }
        net_savings: { type: number }
        transaction_count: { type: integer }
        month: { type: string }

    CategorySpending:
      type: object
      required: [category_id, category_name, category_color, category_icon, amount, percentage]
      properties:
        category_id: { type: integer }
        category_name: { type: string }
        category_color: { type: string }
        category_icon: { type: string }
        amount: { type: number }
        percentage: { type: number }

    MonthlyTrend:
      type: object
      required: [month, income, expenses, net]
      properties:
        month: { type: string }
        income: { type: number }
        expenses: { type: number }
        net: { type: number }

    BudgetVsActual:
      type: object
      required: [category_id, category_name, category_color, budget_limit, actual_spent, remaining, percentage_used]
      properties:
        category_id: { type: integer }
        category_name: { type: string }
        category_color: { type: string }
        category_icon: { type: string }
        budget_limit: { type: number }
        actual_spent: { type: number }
        remaining: { type: number }
        percentage_used: { type: number }