# lib/api-zod/src/generated/types/budgetInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Budget Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface BudgetInput {
  category_id: number;
  month: string;
  /** @minimum 0.01 */
  limit_amount: number;
}